+++ tests/Foundation/Validation/Constraint/UniqueDocumentTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Foundation\Constraint;
+namespace App\Tests\Foundation\Validation\Constraint;
 
-use App\Foundation\Constraint\UniqueDocument;
+use App\Foundation\Validation\Constraint\UniqueDocument;
 use App\Tests\TestCase;
 
 class UniqueDocumentTest extends TestCase
